Disassembled InSitu
This one is interesting:
First of all, it contains two different MBRs, one directly after the
other.
Second, although the two partition tables are identical (as you'd
expect), the rest of the code is different from each other.
Third, I've seen both versions before. The first is identical to Dell,
while the second is nearly identical to GHS1 - I say *nearly* because
they fixed the bug in GHS1!